Extract class parsing into `syntax/class.rs`
